mod: make armor set requirement power mean pvalue to 4 (#447)
* added constant

* wip

* Item Requirement Model uses a constant now

* removed useless lines

* Updated constants.cs and item-service.ts , also changed the line position in constants.json

* ItemRequiment model is not static anymore

* reqmodel

* indentation

* indentation

* format change

* same line in crpgconstants

* readonly

* .

* .

* .

* indentation

+++ src/Module.Server/Helpers/MathHelper.cs
@@ -13,7 +13,7 @@ internal static class MathHelper
         return r;
     }
 
-    public static float GeneralizedMean(int n, float[] numbers)
+    public static float GeneralizedMean(float n, float[] numbers)
     {
         float normN = (float)numbers.Sum(x => Math.Pow(x, n)) / numbers.Length;
         return (float)Math.Pow(normN, 1f / n);
